+++ app/routes/_app+/leaders/$leaderId/_index.tsx
@@ -4,11 +4,10 @@ import { handlerSuccess, handlerError } from "~/.server/reponses";
 import { handlerErrorWithToast } from "~/.server/reponses/handlerError";
 import { handlerSuccessWithToast } from "~/.server/reponses/handlerSuccess";
 import { Service } from "~/.server/services";
+import { permissions } from '~/application';
 
 export const loader: LoaderFunction = async ({ params }) => {
-  
     const { leaderId } = params;
-
     try {
         const leader = await Service.leader.findOne(leaderId);
         return handlerSuccess(200, leader);
@@ -16,9 +15,7 @@ export const loader: LoaderFunction = async ({ params }) => {
         console.log(error);
         return handlerError(error)
     }
-
 }
-
 
 export const action: ActionFunction = async ({ params, request }) => {
     const formData = await request.formData();
@@ -28,22 +25,27 @@ export const action: ActionFunction = async ({ params, request }) => {
     try {
       
         if(data._action === 'update') {
+          await Service.auth.requirePermission(request, permissions.leaders.permissions.update);
           formData.set('folder', data['folder[id]']);
           await Service.leader.updateOne( id, formData );
           return handlerSuccessWithToast('update');
         }
   
         if(data._action === 'delete') {
+          await Service.auth.requirePermission(request, permissions.leaders.permissions.delete);
           await Service.leader.deleteOne(id);
           return handlerSuccessWithToast('delete');
         }
 
+        
         if(data._action === 'subscribe') {
+          await Service.auth.requirePermission(request, permissions.leaders.permissions.active);
           await Service.leader.resubscribe(id, Number(data['folder[id]']));
           return handlerSuccessWithToast('update', 'La lider');
         }
         
         if(data._action === 'unsubscribe') {
+          await Service.auth.requirePermission(request, permissions.leaders.permissions.active);
           await Service.leader.unsubscribe(id, formData);
           return handlerSuccessWithToast('update', 'La lider');
         }
